Добавляет исправления наставника
