+++ index.html
@@ -226,8 +226,6 @@
       </div>
     </section>
   </template>
-  <script type="module" src="js/util.js"></script>
-  <script type="module" src="js/data.js"></script>
   <script type="module" src="js/main.js"></script>
   <script src="js/functions.js"></script>
 </body>
