+++ js/data.js
@@ -1,3 +1,5 @@
+import {getRandomInteger, createRandomUnique, getRandomArrayElement} from './util.js';
+
 const MESSAGES = [
   'Всё отлично!',
   'В целом всё неплохо. Но не всё.',
@@ -8,4 +10,34 @@ const MESSAGES = [
 ];
 const SIMILAR_PUBLICATION_COUNT = 25;
 
-export {MESSAGES, SIMILAR_PUBLICATION_COUNT};
+const generatePostId = createRandomUnique(1, 25);
+const generatePhotoId = createRandomUnique(1, 25);
+const generateCommentId = createRandomUnique(1, 10000);
+
+const randomComments = () => {
+  const randomQuantity = getRandomInteger(0, 30);
+  const comments = [];
+  for (let i = 0; i < randomQuantity; i++) {
+    const comment = {
+      id: generateCommentId(),
+      avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
+      message: getRandomArrayElement(MESSAGES),
+      name: 'Пользователь',
+    };
+    comments.push(comment);
+  }
+
+  return comments;
+};
+
+const createPublication = () => ({
+  id: generatePostId(),
+  url: `photos/${generatePhotoId()}.jpg`,
+  description: 'Описание фотографии',
+  likes: getRandomInteger(15, 200),
+  comments: randomComments(),
+});
+
+const similarPublication = Array.from({length: SIMILAR_PUBLICATION_COUNT}, createPublication);
+
+export {similarPublication };
